refactor(server): split start.ts into separate hub and local entry points
Replace single start.ts with --mode flag with two dedicated entry points
(start-hub.ts, start-local.ts). Each file only imports what it needs,
making the commands self-documenting and avoiding unnecessary dependency
loading. Updated all spec and doc references to match.
