+++ packages/server/package.json
@@ -16,8 +16,12 @@
 	},
 	"license": "AGPL-3.0",
 	"scripts": {
-		"start": "bun src/start.ts",
-		"dev": "bun src/start.ts",
+		"start": "bun src/start-hub.ts",
+"start:hub": "bun src/start-hub.ts",
+"start:local": "bun src/start-local.ts",
+"dev": "bun src/start-hub.ts",
+"dev:hub": "bun src/start-hub.ts",
+"dev:local": "bun src/start-local.ts",
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
