+++ packages/server/src/start-hub.ts
@@ -1,0 +1,42 @@
+/**
+ * Hub server entry point.
+ *
+ * Starts the Epicenter hub — the coordination server that provides:
+ * - Sync relay (primary) — all devices sync through the hub
+ * - AI streaming — all providers via SSE
+ * - AI proxy — env var API keys, never leave the hub
+ * - Better Auth — session-based authentication
+ *
+ * Usage:
+ *   bun packages/server/src/start-hub.ts
+ *   PORT=4000 bun packages/server/src/start-hub.ts
+ */
+
+import { createHubServer } from './hub';
+
+const port = Number.parseInt(process.env.PORT ?? '3913', 10);
+
+const server = createHubServer({
+	port,
+	sync: {
+		onRoomCreated: (roomId) => console.log(`[Sync] Room created: ${roomId}`),
+		onRoomEvicted: (roomId) => console.log(`[Sync] Room evicted: ${roomId}`),
+	},
+});
+
+server.start();
+
+console.log(`Epicenter HUB server running on http://localhost:${port}`);
+console.log(`  Sync:    ws://localhost:${port}/rooms/{room}`);
+console.log(`  AI:      POST http://localhost:${port}/ai/chat`);
+
+process.on('SIGINT', async () => {
+	console.log('\nShutting down...');
+	await server.stop();
+	process.exit(0);
+});
+
+process.on('SIGTERM', async () => {
+	await server.stop();
+	process.exit(0);
+});
