+++ specs/20260221T190252-ai-chat-tab.md
@@ -269,7 +269,7 @@ Max Concurrent: 3 (Wave 1)
     Tool: Bash (curl)
     Preconditions: Server running with OPENAI_API_KEY=sk-test-invalid-key env var
     Steps:
-      1. Start server: OPENAI_API_KEY=sk-test-key bun packages/server/src/start.ts &
+      1. Start server: OPENAI_API_KEY=sk-test-key bun packages/server/src/start-hub.ts &
       2. curl POST to /ai/chat with provider=openai, no x-provider-api-key header
       3. Assert: HTTP status is NOT 401
     Expected Result: Status 200 (streaming) or 502 (provider error from invalid key) — NOT 401
@@ -641,7 +641,7 @@ cd apps/tab-manager && bun run typecheck
 cd apps/tab-manager && bun run build
 
 # Server env var API key works
-OPENAI_API_KEY=sk-test bun packages/server/src/start.ts &
+OPENAI_API_KEY=sk-test bun packages/server/src/start-hub.ts &
 curl -s -X POST http://localhost:3913/ai/chat \
   -H 'Content-Type: application/json' \
   -d '{"messages":[{"role":"user","content":[{"type":"text","content":"say hi"}]}],"provider":"openai","model":"gpt-4o-mini"}' \
